+++ src/shader.rs
@@ -1,0 +1,9 @@
+mod light;
+mod camera;
+mod cel;
+mod outline;
+
+pub use light::*;
+pub use camera::*;
+pub use cel::*;
+pub use outline::*;
