--- src/shaders.rs
@@ -1,5 +1,0 @@
-mod cel;
-mod outline;
-
-pub use cel::*;
-pub use outline::*;
