+++ src/tasks.rs
@@ -1,0 +1,7 @@
+//! The various tasks that spritec can execute
+
+mod pose;
+mod spritesheet;
+
+pub use pose::*;
+pub use spritesheet::*;
